feat(pathfinder/sync/l2): calculate 0.13.2 commitments for pre-0.13.2 blocks
For pre-0.13.2 blocks we actually have to re-compute the transaction
and event commitments: after we've verified that the block hash is
correct we no longer need the legacy commitments. The P2P protocol
requires that all commitments in block headers are the 0.13.2 variants
for legacy blocks, so we need to have those in our storage.