fix(pathfinder/block_hash): fix transaction commitment calculation for empty signatures
Turns out our 0.13.2 transaction commitment calculation algorithm was
slightly different compared to the one used by the sequencer: for
transaction types that _do_ have a signature field but it's empty we
are required to use `[0]` as the signature when calculating the
commitment leaf value.

This issue was triggered in Sepolia testnet block 218484 that contains
an Invoke v1 transaction with an empty signature field.

Closes: #2282