feat(executor): add structured transaction execution errors
In JSON-RPC 0.8.0 we're expected to return the error in a structured
format for reverted and failing transactions.

Unforunately `blockifier` 0.8.0-rc.3 does not directly expose that
information: attributes of the `ErrorStack` type in blockifier are
private, plus the revert error only comes out of blockifier as a
string containg the formatted stack trace.

These are expected to become available in the blockifier version that
will be used for Starknet 0.13.3 -- but until that  becomes available
we're forced to use our blockifier fork fixing these API issues.

This change adds error stack information to our executor API so that
we _do_ expose the necessary information to the JSON-RPC implementation.