fix(storage): work around in-memory DB locking issues
We're using shared cache mode with our in-memory DB to allow multiple
connections from within the same process. This means that in contrast to
a file-based DB we immediately get locking errors in case of concurrent
writes -- a pool size of one avoids this.

There are certain tests that _do_ need multiple connections to the
in-memory DB though so we add a constructor that allows specifying the
size of the pool we require.