fix(rpc/v08): fix spec compliance in `starknet_getBlockWithReceipts` response
The OpenRPC specification requires sending an array of {transaction,
receipt} objects.

Note that the transaction in this schema is #/components/schemas/TXN,
which does not contain a transaction_hash property. We're always
serializing transactions with the transaction_hash property present
which makes our responses to starknet_getBlockWithReceipts not compliant
with the specification.

Closes: #2420