fix(storage/block): fix reorg storage problem
As part of the still-in-development "aggregate bloom filters feature we
have added a SQL statement to `purge_block()` that causes issues
during a reorg because it references a non-existent table when compiling
Pathfinder _without_ the "aggregate_bloom" feature.

This change fixes that by gating the SQL statement behind the feature.