feat(executor): conditionally enable cairo-native
The new `cairo-native` feature enables the use of Cairo native in our
executor code. Currently all Sierra classes are compiled into native
at a significant runtime cost (the cost of compiling using MLIR and LLVM,
plus additionally linking the resulting object file into a shared object).

To make this compile the following environment variables need to be set
to the root of the LLVM 19.1 installation: MLIR_SYS_190_PREFIX,
LLVM_SYS_191_PREFIX, and TABLEGEN_190_PREFIX.

The `CAIRO_NATIVE_RUNTIME_LIBRARY` environment variable must be set to the
path of `libcairo_native_runtime.a`. Accidentally things just work if it
points to an empty file: in that case the missing symbols are resolved
using those in the Pathfinder executable (which is linked with the Cairo
Native runtime).