feat(rpc/v06): re-add block getters
For `starknet_getBlockWithTxHashes` and `starknet_getBlockWithTxs` the
only difference compared to v07 is that the block header and pending
block header types are missing L1 data gas and L1 DA mode fields.

This change also adds some rudimentary tests for testing the serialization
of the block header for all the different JSON-RPC versions we're
supporting.

Closes: #2554