feat(rpc/test_setup): use OpenZeppelin-based dummy account
Instead of our old Cairo 0 dummy account, we should be using a modern,
Cairo 1 account compiled with Cairo 2.10 (to get a Sierra 1.7 class).

This is required in order to be able to utilize Sierra gas measurements
during fee computation for Starknet 0.13.4.

The account class added is based on OpenZeppelin Cairo contracts v0.20.0
"preset" class with the `is_valid_stark_signature` function patched to
always return true.