fix(executor): derive max L2 fee from balance
Blockifier has a pre-check for `AccountTransaction` that checks that the
account has sufficient balance to cover all committed resource bounds.

Because of this `starknet_simulateTransaction` fails with "insufficient
balance" when we're initially running the transaction with L2 gas max
amount set to u64::MAX.

We should derive the max L2 gas amount so that the balance of the
account still covers all the resource bounds.