refactor(rpc): remove redundant transaction serialization
We had two transaction serialization implementations in the RPC module.
One of them in the `dto` module, used by almost all methods, and one in
`types` that was exclusively used by `starknet_getTransactionByBlockIdAndIndex`.

This change removes the implementation in `types` and converts the only
user to use the one in `dto`.