fix(executor): fix fee estimation for the all-zero-bounds case
Fee estimation is usually done by specifying zeroes as bounds for the
transaction (since the values to use are not known yet).

However, our code was unconditionally doing a check that failed the
estimation if the actual L2 gas limit we have determined was above
the maximum gas limit in the transaction resource bounds.

This change modifies that check to only fail if we're not skipping
fee transfer (ie. only for `starknet_simulateTransactions` with
the `SKIP_FEE_TRANSFER` flag not set).