fix(rpc/subscription): send notifications for `latest` block
When subscribing with the `latest` block, we still have to do catch-up.
The JSON-RPC specification says that the starting point for the
subscription should be the block specified by `block_id` in the
subscription request. This means that we have to send a notification for
the `latest` block, irrespective if it was requested explicitly or
implicitly and if we used the `latest` label or its specific block number
or hash.

Closes #2642