feat(rpc/subscribe_events): send matching events from the pending block
This PR adds the ability to send events matching the filter from the
pending block to the client right as they are included in the pending block.
This minimizes the latency it takes for users to receive relevant updates.

One side effect is that since the pending block has no block
hash, we send out emitted events without the block hash as a context.
Since almost all events show up first in the pending block, in practice
this means that we are no longer sending block hashes for the emitted
events...