fix(gateway-types): add missing `failed` field to transaction trace function invocation
Starknet 0.13.4 has added a new `failed` field to function invocations,
causing parsing to fail. This change adds the new field as optional and
also uses that value when converting gateway function invocations to
JSON-RPC function invocations (`is_reverted` property).

This latter conversion is not really used though, as we're not falling
back to the feeder gateway for traces for Starknet 0.13.4 blocks.

Closes: #2647