fix(ethereum): gracefully handle L1 finalized block fetch failure
If fetching the L1 finalized block fails due to a network error, we now
exit from the poll task (instead of leaking it and leaving it retrying
indefinitely). This causes the `finalized_block_rx` channel to be closed,
which we now also handle in the main loop of `sync_and_listen()` by
exiting and returning an error. This in turn leads to sync restarting the
L1 sync process.