replace `Storage` with `Transaction` in state reader
A pruning feature can be enabled in Pathfinder to delete block data more
than N blocks back. Since the state reader provides data to multiple
transactions in one batch, it is unsafe to create a new database
transaction for each Starknet transaction because some of the data might
be pruned for one transaction but not pruned for the others, leading to
inconsistent data.