fix(executor): simulate should return reverted transaction if L2 cap is too low
After finding the minimum L2 gas, we are re-running the transaction in case
the initial L2 gas value in the transaction is too low. This is done so
that simulate can return the proper error / reverted transaction.

There were two issues with our code:
1. We were always executing using `ExecutionBehaviorOnRevert::Fail`, which
    means that we were not able to return the reverted transaction (but
    return an error instead).
2. We were using an L2 gas vale of zero instead of the initial L2 gas
    during execution, which potentially lead to a different error.

This change addresses these issues and also fixes an imprecision. We
have been using the transaction receipt in `TransactionExecutionInfo` to
return the proper minimun L2 gas value to the caller in
find_l2_gas_limit_and_execute_transaction(). However, that receipt is
also used to produce the `execution_resources` field in the transaction
trace, which in turn should contain the _actual_ L2 gas consumption.

find_l2_gas_limit_and_execute_transaction() now returns both: the
transaction execution info should contain the _actual_ resource usage,
and we return the gas vector that should be used to compute the fee
estimate (using the minimum L2 gas) explicitly.