fix(rpc/trace): don't fall back to fetching traces from feeder gateway for unsupported versions
When doing transaction traces we have a check that checks if the Starknet
version is supported. If it's less than the earliest version we support
for execution, then we try and fall back to the feeder gateway before
even trying execution.

We also had a check that if a `TransactionExecutionError` was returned
during tracing, we would fall back to the feeder gateway. This is not needed
and can be removed, since at that point we can be sure that the version
is supported and the feeder gateway will just return an error (it doesn't
actually have traces from 0.13.1.1, our earliest supported version).

This change removes this fallback, which in turn was masking the error
message that was generated during execution: the feeder gateway error
was transformed into an `InternalError` and no details were included
in the JSON-RPC response.

Now we will return the error message from the execution, which is more
useful for the end user.