fix(storage/events): don't return non-contiguous aggregate filters in `load_event_filter_range()`
There's a bug in the processing logic when we combine `max_event_filters_to_load`
in the SQLite query and cached filters that are within the block range.

Due to simply appending the cached aggregate filters to the set of filters
returned it's possible that there's a hole in the block range covered by
the filters. In case there are more than `max_event_filters_to_load`
filters in the database, we stop loading filters when we reach the limit.
However, the cached filters that are appended to the result set may not
be contiguous with the last filter in the result set.

This breaks the assumption that the filters returned by `load_event_filter_range()`
are contiguous, which causes `events()` to just skip evaluating
ranges of blocks that are not covered by the filters.

Closes #2709