fix(rpc): handle missing call_type field in feeder gateway block trace responses
For pre-0.9 Starknet blocks the feeder gateway returns a block trace response
that is missing the call_type, class_hash, entry_point_type and selector fields
for function invocations.

This commit updates the code to handle this case gracefully by not including
these fields in the JSON-RPC response. Technically this is not compliant
with the specification but is _still_ less misleading than returning
bogus values.