fix(executor): fix DEPLOY_ACCOUNT transaction estimation
Turns out that we have two issues with DEPLOY_ACCOUNT estimation, both
triggered by Braavos account deployments.

The first one is that `l2_gas_accounting_enabled()` does not work for
all DEPLOY_ACCOUNT transactions. We were trying to fetch the class hash
for the sender of the transaction, but for DEPLOY_ACCOUNT that is the
address of the account that is being deployed, so no class hash is
present yet. Because of this we were always running DEPLOY_ACCOUNT
transactions with L2 gas accounting disabled.

Even if we had the class hash, we would still fail for Braavos accounts.
The trick is that Braavos accounds are deployed in a special way: they
are always using the same class hash, and the constructor immediately
replaces the deployed class hash with a new one that is received as
part of the "signature". This leads to the validate entry point actually
being executed in a completely different class than the one that is
present in the `class_hash` field of the transaction. This completely
breaks our L2 gas accounting detection logic.

This commit changes the detection logic so that we always run DEPLOY_ACCOUNT
with L2 gas accounting enabled. (This incurs a performance penalty though,
because we now execute the transaction at least twice.)

The second issue is that when running the L2 gas accounting aware
estimation logic, we derive the max L2 gas bound from the balance of the
sender account. However, for account deployments it's perfectly possible
that the account balance is still zero at the time of fee estimation.

This commit special cases DEPLOY_ACCOUNT transactions and just uses an
upper limit for the L2 gas bound from versioned constants.

Closes #2734