remove foreign key constraints to canonical_blocks
In order to simplify blockchain pruning logic, some of the foreign key
constraints to the canonical_blocks table had to be removed. They were
there to make purging blocks during reorgs easier but were not of much
other use. With the foreign keys gone (and some slight changes), the
canonical_blocks table did not serve any other purpose so it has been
dropped.