feat(pathfinder/sync): implement polling of pre-confirmed blocks
With Starknet 0.14.0, the pending block is gone, and we have the
pre-confirmed block instead. This new concept is similar, but the
contents are of a different level of finality.

For JSON-RPC 0.9, which has the new transaction status values, we can
represent the new reality just fine. However, for previous JSON-RPC
versions we'll need to ignore the contents of the pre-confirmed block.

This commit implements the polling logic and basic data structures to
support the pre-confirmed block. The logic is similar to the pending
block polling, but it uses the new `get_pre_confirmed_block` feeder
gateway endpoint.

The transition from Starknet 0.13.x to 0.14.0 is implemented the
following way:

- We start polling the pending block initially, as before. In case the
    Starknet version is 0.14.0 or higher, we switch to polling the
    pre-confirmed block instead.

- Polling the pre-confirmed block is done in a loop, using the latest
    block number + 1 as the parameter for the feeder gateway endpoint.

The RPC crate now has a data type for storing pre-confirmed block
information and we do a conversion whenever we receive a new pre-confirmed
block from the sync logic. This is needed so that we can store the
candidate transactions (transactions that have not been executed yet)
separately from pre-confirmed transactions.