feat(pathfinder): poll latest and pending/pre-confirmed block every second
With the drastic decrease of block time in Starknet 0.14.0 we need to
lower the polling interval to 1 second to ensure that we are up-to-date
with the current state of the network.

Since the pre-confirmed block can only be polled using an explicit block
number the latest block information is now a limitation to the pre-confirmed
polling too.

This change decreases the default head polling interval to one second
and uses the same interval for the polling of the pre-confirmed block.