fix(executor): correct max L2 limit in L2 gas estimation logic
Previously, we have introduced a change that sets the maximum L2 gas limit
for the initial transaction run to the maximal value that the account
balance allows. This is done because blockifier _does_ check that the
committed resources are covered by the account balance.

However, that check is done only if the `charge_fee` flag is set (which
is the case for `starknet_simulateTransactions` only, unless
`SKIP_FEE_CHARGE` is set).

This means that for `starknet_estimateFee` we cannot estimate
transactions for accounts that are not yet funded properly, even
though blockifier would allow it.

This commit changes the logic to use the balance-dependent limit only
when the `charge_fee` flag is set, and uses blockifier's default
limit for per-transaction L2 gas otherwise.