fix(rpc/pending): properly set parent state commitment in state update
Both the empty state update we're constructing and the pre-committed
state update needs to have the parent state commitment field set to the
state commitment of the parent block.

This is especially important for the pre-committed state update, as the
parent state commitment there is the _only_ way to identify which
parent block the state update is for (the block number is not
included in the response for `starknet_getStateUpdate`).