feat(rpc/v09): update INVALID_TRANSACTION_NONCE error with additional data
And make sure we return the INVALID_TRANSACTION_NONCE error in case
the gateway is returning an invalid nonce error.

This is slightly complicated by the fact that the gateway is returning
a "validation failure" error instead of the "invalid nonce" error, so
the only way for us to detect this is to check the error message
for the presence of the "invalid nonce" string.