fix(rpc): fix pre-confirmed subscription data consistency
In the sync code where we were emitting the pre-confirmed updates, we
did _not_ check if the pre-confirmed block is consistent with the latest
block we have in the database. Having an inconsisntent pre-confirmed
block update is not a problem for normal JSON-RPC methods, because those
methods always check the consistency of the pre-confirmed data before
using it.

However, subscription methods using pre-confirmed data are different:
they do not check consistency, just use the pre-confirmed updates as is.

Because of this we _do_ need to ensure that pre-confirmed block updates
are emitted only if the pre-confirmed block is consistent with `latest`.