chore(cargo): add a `ci-dev` Cargo profile
This profile disables incremental compilation (not useful on CI) and
enables minimal debug information only (so that we have line number
information).

We also remove the Cargo build job limit of just two jobs to speed up
compilation. It seems that we run out of memory during _linking_ so
the build job limit does not really fix the issue of timeouts but still
slows down our builds.