fix(rpc/method/get_events): fix `starknet_getEvents` for empty blocks
When processing a query with a block range that involves empty blocks
(blocks with zero transactions / events) `starknet_getEvents` would
stop scanning at the first empty block.

This is bad, because it doesn't even return a continuation token so
users would have no way of knowing that there are more events.

The issue is that `query_events_by_block()` is returning an `Option`,
and `None` is being interpreted as "no block found",
which causes the scan to stop prematurely when encountering empty blocks.

To fix this, we need to ensure that we handle no events correctly in
`events_for_block()` and just return an empty vector of events instead
of returning `None`.