fix(docker): don't reinstall Rust toolchain
Ever since we've added `rust-toolchain.toml` to the repository, Docker
builds do re-download the toolchain even though we should just be using
the one already installed in the image.

This commit updates the Dockerfile to exclude the `rust-toolchain.toml`
file from the build context, so that the toolchain is not reinstalled.