fix(sync): bump block cache size to 10k blocks
Pathfinder can only do reorgs within the capacity of the block cache.

Unfortunately the recent L2 reorg on Starknet Mainnet was larger than
the previous cache size of 1k blocks, which caused the reorg attempt to
fail and Pathfinder was stuck in a loop trying to do the reorg again
and again.