fix(pathfinder/sync/l2): properly check for reorgs during bulk sync
When performing a bulk sync, we need to validate that the blocks we are
receiving are indeed a continuation of our current chain. This is done
after performing all other validations.

Note that we do _not_ handle reorgs during bulk sync, but simply give
up bulk syncing and revert back to the block-by-block sync algorithm
that then handles the reorg itself.