refactor: move feeder-gateway example to its own crate
The example did not work since the introduction of aggregate Bloom
filters.

The issue was that the pathfinder crate enables the
`small_aggregate_filters` feature for tests, but that also affects the
build configuration for examples. `small_aggregate_filters` is
incompatible with production use, so the feeder-gateway example would
just fail with with an error immediately upon startup.