fix(feeder-gateway): re-calculate event and transaction commitments
For pre-0.13.2 blocks we're storing 0.13.2 transaction and event
commitment values in storage (so that we don't have to re-calculate
these when serving a P2P request).

However, on the feeder gateway interface simply returning these values
does not work, and leads to commitment mismatch errors on clients.

This fix changes the feeder-gateway code to re-calculate these
commitments using the legacy method for pre-0.13.2 blocks.