feat(rpc): implement feeder gateway fallback for non-reexecutable mainnet range
Starknet mainnet now has a block range, that -- due to a sequencer bug --
is not re-executable (re-execution leads to different results).

Starkware has re-opened the feeder gateway endpoint for fetching
transaction traces for the affected range, so we can now fall back to
fetching the results from the feeder gateway for that range.