fix(rpc/v0_8/subscribe_pending_transactions): avoid busy loop with Starknet 0.14.0
After the Starknet 0.14.0 update, the pending block is no more. Some
code has been added to the `subscribe_pendingTransactions` handler to
ignore pre-confirmed block updates, but unfortunately it has resulted
in a busy loop, as we never actually wait for a new pending update
before checking the block again.

This change fixes that loop and makes sure we actually wait for an
update before performing the check again.