fix(pathfinder/sync/l2): remove storage access from reorg-detection logic
The L2 sync "producer" code -- that is, the task which is polling the
feeder gateway and producing sync events -- should work more-or-less
independently of the storage.

This is actually the only way it can work reliably, because the actual
storage changes are performed asynchronously. Reorg detection simply
cannot assume that storage state matches the state of the "producer".

This change removes a storage block hash lookup from the reorg
detection code and uses the "block cache" instead, which is an
in-memory representation of the last 10k blocks produced by the L2
sync "producer" task.