fix(feeder-gateway): fix reorg simulation
Now that we have more precise logic for reorg detection in Pathfinder,
some changes were required to the feeder gateway reorg simulation logic.

Pathfinder now waits for the feeder gateway to serve a non-matching
"latest" block before detecting a reorg. We now force this by
forcibly changing the block hash in block headers for blocks in the
reorged-away range.