fix(storage/schema/revision_0074): update block range when rebuilding filter
When rebuilding the running event filter, ensure that the block range
(from_block to to_block) is also updated to reflect the range of blocks
covered by the rebuilt filter.

The database might have a completely outdated running event filter, so
it's important to ensure that the correct block range is set.