fix(sync/l2): retry fetching signature until we succeed
There is a race condition with the current setup of Starknet feeder
gateway: we've observed that fetching the signature sometimes fails
even after fetching the block itself succeeds. We currently fail on such
errors and exit the L2 sync task, which is not ideal.

This change modifies the logic to retry fetching the signature until
it succeeds.