fix(sync): lower L2 sync restart delay to 5s
The L2 sync task can exit on a couple of conditions, including receiving
an unexpected Starknet error from the feeder gateway while fetching
data. In such cases, we want to retry quickly, so we lower the restart
delay from 60 seconds to 5 seconds.

The 60 seconds delay made sense originally when we had multi-minute
block times, but with the current block times it causes the node to fall
behind significantly.