fix(rpc): fix `starknet_simulateTransaction` for `DEPLOY_ACCOUNT`
When simulating `DEPLOY_ACCOUNT` transactions without the
`SKIP_FEE_CHARGE` flag execution _does_ verify that the balance of the
account is sufficient to cover the resource bounds in the transaction.

Unfortunately we did a change a few months ago that now breaks this:
`DEPLOY_ACCOUNT` transactions are always executed using the initial gas
setting from versioned constants in case `SKIP_FEE_CHARGE` is not set.
This means that the balance check is now always performed against these
initial gas values, which are quite high, and thus the balance check
often fails.

The fix is to not use the initial gas values for `DEPLOY_ACCOUNT`
transactions, but instead use the max L2 gas value calculated from the
account balance, just as we do for other transaction types.

Note that this change only affects simulation without `SKIP_FEE_CHARGE`.
For estimation (which is implicitly with `SKIP_FEE_CHARGE`) we still use
the initial gas values because the goal there is to find out how much
STRK should be pre-funded in the account address so that the deployment
succeeds.