feat(storage): migrate CASM definitions to include block number
CASM hashes are now not unique by Sierra class hash alone, but by the combination
of class hash and block number. This migration updates the database schema to
reflect this change.

CASM class hashes are no longer stored in the `casm_definitions` table, but are
split out to their own `casm_class_hashes` table. In this table (class_hash, block_number)
is a unique key.