feat(storage): add pre-computed CASM v2 hashes
Blockifier on Starknet 0.14.1 requires being able to look up Blake2 CASM
class hashes for all CASM hashes that are executed.

Unfortunately, computing these Blake2 hashes on-the-fly is quite
expensive, so we add a pre-computed mapping from class hashes to Blake2
CASM class hashes to the storage layer.

This migration step uses two binary files containing the pre-computed
mappings for a subset of mainnet and Sepolia testnet respectively. These
files were generated using the `recompute_casm_class_hashes` example.
The files contain a mapping of (ClassHash, CasmHash) pairs and are used
to avoid parsing the CASM class definitions and computing the Blake2
CASM class hash for all classes for which we already have the result
for.