fix(executor/state_reader): fix CASM hash v2 computation
In case we fall back to computing the CASM hash v2 from the compiled
class we should use `class.hash(&HashVersion::V2)` instead of the
utility function.

If the Sierra class is declared within the batch of transactions being
executed, we do _not_ have the CASM definition available in storage yet
so we need to compute the CASM hash v2 from the compiled class passed as
an argument.