fix(Dockerfile): fix libc symlink creation
Cairo Native compilation requires a working linker and access to the
libc.so file (so that the `-lc` linker argument works).

The symlink creation in the Dockerfile was not working as intended
because it was using a POSIX shell construct that is not supported by
the default `/bin/sh` shell in Debian-based images (which is `dash`).