refactor(state): update state only using sync
- Instead of a catch-up sync task that is kicked off whenever a node
falls behind the rest of the consensus network (and block updates being
done by consensus otherwise), introduce a consensus-aware sync task that
does all L1/L2 state updates.
- The new task is consensus-aware inasmuch as it first requests the
block that it is about to sync next from the consensus task and, if it
receives the block back, uses it for the L2 update instead of
downloading from a data source (FGW or sync P2P network, though this
commit it is limited to FGW).
- For this to work, the consensus task needs to store a cache of most
recently finalized blocks, which it currently does in a separate
database used for fake proposals.