fix(rpc/v10): Omit pre-0.13.2 tx and event commitments
We store Starknet 0.13.2-style transaction and event commitments for all
pre-0.13.2 blocks in order to support the P2P sync protocol. However,
these commitments are not the same as the original commitments for these
blocks, which may lead to confusion.

To avoid this, we return zeroed commitments for pre-0.13.2 blocks in the
RPC responses.